get rid of sync http mon
