+++ ydb/core/mon/mon.h
@@ -1,29 +1,26 @@
 #pragma once
 
-#include <library/cpp/json/writer/json_value.h>
+#include <future>
 #include <library/cpp/monlib/service/monservice.h>
 #include <library/cpp/monlib/dynamic_counters/counters.h>
+#include <library/cpp/monlib/service/pages/index_mon_page.h>
 #include <library/cpp/monlib/service/pages/resources/css_mon_page.h>
 #include <library/cpp/monlib/service/pages/resources/fonts_mon_page.h>
 #include <library/cpp/monlib/service/pages/resources/js_mon_page.h>
 #include <library/cpp/monlib/service/pages/tablesorter/css_mon_page.h>
 #include <library/cpp/monlib/service/pages/tablesorter/js_mon_page.h>
 
-#include <ydb/library/actors/core/actor.h>
 #include <ydb/library/actors/core/mon.h>
+#include <ydb/library/actors/http/http.h>
 #include <yql/essentials/public/issue/yql_issue.h>
 #include <ydb/public/sdk/cpp/client/ydb_types/status/status.h>
 
+#include "mon.h"
+
 namespace NActors {
 
-IEventHandle* SelectAuthorizationScheme(const NActors::TActorId& owner, NMonitoring::IMonHttpRequest& request);
-IEventHandle* GetAuthorizeTicketResult(const NActors::TActorId& owner);
-
-void MakeJsonErrorReply(NJson::TJsonValue& jsonResponse, TString& message, const NYql::TIssues& issues, NYdb::EStatus status);
 void MakeJsonErrorReply(NJson::TJsonValue& jsonResponse, TString& message, const NYdb::TStatus& status);
-
-class TActorSystem;
-struct TActorId;
+void MakeJsonErrorReply(NJson::TJsonValue& jsonResponse, TString& message, const NYql::TIssues& issues, NYdb::EStatus status);
 
 class TMon {
 public:
@@ -45,12 +42,14 @@ public:
         TDuration InactivityTimeout = TDuration::Minutes(2);
     };
 
+    TMon(TConfig config);
     virtual ~TMon() = default;
-    virtual void Start(TActorSystem* actorSystem = {}) = 0;
-    virtual void Stop() = 0;
-    virtual void Register(NMonitoring::IMonPage* page) = 0;
 
-    virtual NMonitoring::TIndexMonPage* RegisterIndexPage(const TString& path, const TString& title) = 0;
+    std::future<void> Start(TActorSystem* actorSystem); // signals when monitoring is ready
+    void Stop();
+
+    void Register(NMonitoring::IMonPage* page);
+    NMonitoring::TIndexMonPage* RegisterIndexPage(const TString& path, const TString& title);
 
     struct TRegisterActorPageFields {
         TString Title;
@@ -65,12 +64,32 @@ public:
         TString MonServiceName = "utils";
     };
 
-    virtual NMonitoring::IMonPage* RegisterActorPage(TRegisterActorPageFields fields) = 0;
+    NMonitoring::IMonPage* RegisterActorPage(TRegisterActorPageFields fields);
     NMonitoring::IMonPage* RegisterActorPage(NMonitoring::TIndexMonPage* index, const TString& relPath,
         const TString& title, bool preTag, TActorSystem* actorSystem, const TActorId& actorId, bool useAuth = true, bool sortPages = true);
-    virtual NMonitoring::IMonPage* RegisterCountersPage(const TString& path, const TString& title, TIntrusivePtr<::NMonitoring::TDynamicCounters> counters) = 0;
-    virtual NMonitoring::IMonPage* FindPage(const TString& relPath) = 0;
-    virtual void RegisterHandler(const TString& path, const TActorId& handler) = 0;
+    NMonitoring::IMonPage* RegisterCountersPage(const TString& path, const TString& title, TIntrusivePtr<::NMonitoring::TDynamicCounters> counters);
+    NMonitoring::IMonPage* FindPage(const TString& relPath);
+    void RegisterHandler(const TString& path, const TActorId& handler);
+
+protected:
+    TConfig Config;
+    TIntrusivePtr<NMonitoring::TIndexMonPage> IndexMonPage;
+    TActorSystem* ActorSystem = {};
+    TActorId HttpProxyActorId;
+    TActorId HttpMonServiceActorId;
+    TActorId NodeProxyServiceActorId;
+
+    struct TActorMonPageInfo {
+        NMonitoring::TMonPagePtr Page;
+        TActorId Handler;
+        TString Path;
+    };
+
+    TMutex Mutex;
+    std::vector<TActorMonPageInfo> ActorMonPages;
+    THashMap<TString, TActorId> ActorServices;
+
+    void RegisterActorMonPage(const TActorMonPageInfo& pageInfo);
 };
 
 } // NActors
