+++ ydb/core/mon/ya.make
@@ -1,12 +1,8 @@
 LIBRARY()
 
 SRCS(
-    async_http_mon.cpp
-    async_http_mon.h
     mon.cpp
     mon.h
-    sync_http_mon.cpp
-    sync_http_mon.h
     crossref.cpp
     crossref.h
 )
