+++ ydb/core/testlib/actors/test_runtime.cpp
@@ -4,8 +4,7 @@
 #include <ydb/core/base/blobstorage.h>
 #include <ydb/core/base/counters.h>
 #include <ydb/core/base/pool_stats_collector.h>
-#include <ydb/core/mon/sync_http_mon.h>
-#include <ydb/core/mon/async_http_mon.h>
+#include <ydb/core/mon/mon.h>
 #include <ydb/core/mon_alloc/profiler.h>
 #include <ydb/core/grpc_services/grpc_helper.h>
 #include <ydb/core/tablet/tablet_impl.h>
@@ -193,19 +192,11 @@ namespace NActors {
 
             if (NeedMonitoring && !SingleSysEnv) {
                 ui16 port = MonitoringPortOffset ? MonitoringPortOffset + nodeIndex : GetPortManager().GetPort();
-                if (MonitoringTypeAsync) {
-                    node->Mon.Reset(new NActors::TAsyncHttpMon({
-                        .Port = port,
-                        .Threads = 10,
-                        .Title = "KIKIMR monitoring"
-                    }));
-                } else {
-                    node->Mon.Reset(new NActors::TSyncHttpMon({
-                        .Port = port,
-                        .Threads = 10,
-                        .Title = "KIKIMR monitoring"
-                    }));
-                }
+                node->Mon.Reset(new NActors::TMon({
+                    .Port = port,
+                    .Threads = 10,
+                    .Title = "KIKIMR monitoring"
+                }));
                 nodeAppData->Mon = node->Mon.Get();
                 node->Mon->RegisterCountersPage("counters", "Counters", node->DynamicCounters);
                 auto actorsMonPage = node->Mon->RegisterIndexPage("actors", "Actors");
